Change i32 => i64 for Int

+++ rust/tests/common.rs
@@ -39,7 +39,7 @@ use alga::general::TwoSidedInverse;
 
 /// An integer value
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub struct Int(pub i32);
+pub struct Int(pub i64);
 
 /// Binary operator for calculating the arithmetic sum.
 /// Has the following properties:
@@ -94,7 +94,7 @@ impl Operator for Max {
 
 impl Identity<Max> for Int {
     fn identity() -> Int {
-        Int(std::i32::MIN)
+        Int(std::i64::MIN)
     }
 }
